ENG-19: Extract carousel into separate component
