+++ types/carouselData.ts
@@ -1,0 +1,7 @@
+import { ImageSourcePropType } from "react-native";
+
+export type CarouselData = {
+  title: string;
+  image: ImageSourcePropType;
+  description: string;
+};
